refactor: moved items dialogs to own files
